+++ frontend/src/pages/Service.vue
@@ -289,7 +289,7 @@ function getData(vm) {
     const api = new ServiceApi();
 
     vm.service = null;
-    api.serviceGet(getServiceId(vm), (error, data) => {
+    api.serviceGet(getServiceId(vm), {allowAdmin: 'false'}, (error, data) => {
         if (!error) {
             vm.service = data;
         }