+++ frontend/src/pages/SystemSettings--Features.vue
@@ -11,7 +11,8 @@
                 Check this if the API for applications should not return groups
                 for a player account if one or more of its characters have an invalid token
                 (no token or tokens without any scopes count as invalid), "managed" accounts
-                are excluded from this.
+                are excluded from this.<br>
+                This also affects groups given to Neucore plugins.
             </label>
         </div>
         <label class="mt-2 display-block">